+++ src/common/config.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-undef */
 export default {
-  worker: {
+  email: {
     defaultFromEmail: process.env.DEFAULT_FROM_EMAIL || 'notifications@nkinstallations.com',
     defaultFromName: process.env.DEFAULT_FROM_NAME || 'NK Notifications',
     siteUrl: process.env.SITE_URL || 'http://localhost:7700',
@@ -15,9 +15,9 @@ export default {
     secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
     region: process.env.AWS_REGION,
     convertedImagesBucket: process.env.CONVERTED_IMAGES_BUCKET,
-    originalImagesBucket: process.env.ORIGINAL_IMAGES_BUCKET,
+    originalImagesBucket: process.env.ORIGINAL_IMAGES_BUCKET || '',
     sqsUrl: process.env.SQS_URL || '',
-    sqsBucket: process.env.SQS_BUCKET,
+    sqsBucket: process.env.SQS_BUCKET || '',
   },
   port: process.env.PORT || 7702,
   env: process.env.NODE_ENV || 'development',