+++ src/libraries/image-convert/index.ts
@@ -1,0 +1,3 @@
+import convertImage from './functions/convertImage.js'
+
+export {convertImage}