+++ src/libraries/storage/index.ts
@@ -1,0 +1,5 @@
+import deleteObject from './functions/deleteObject.js'
+import getObject from './functions/getObject.js'
+import listObjects from './functions/listObjects.js'
+
+export {deleteObject, getObject, listObjects}