+++ src/types/image.ts
@@ -1,0 +1,3 @@
+export interface ImageMessage {
+  filename: string[]
+}