+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.2.0](https://github.com/CandeeGenerations/nk-consumer/compare/v1.1.0...v1.2.0) (2024-09-03)
+
+
+### Features
+
+* finished moving image processing to the consumer ([867b059](https://github.com/CandeeGenerations/nk-consumer/commit/867b05933dfbcb37f7fe11da27385bcd6032b84b))
+* working on adding image processing support ([a260d64](https://github.com/CandeeGenerations/nk-consumer/commit/a260d64cbf62dd70641f4374c12f14ff3af85f9f))
+
 ## 1.1.0 (2024-08-30)